+++ src/PokeAByte.Domain/Interfaces/IPokeAByteInstance.cs
@@ -15,6 +15,10 @@ public interface IPokeAByteInstance : IAsyncDisposable
 
     bool? GetModuleFunctionResult(string function, IPokeAByteProperty property);
     public event InstanceProcessingAbort? OnProcessingAbort;
+    public Task FreezeProperty(IPokeAByteProperty target, byte[] bytesFrozen);
+    public Task UnfreezeProperty(IPokeAByteProperty target);
+    public Task WriteValue(IPokeAByteProperty target, string value, bool? freeze);
+    public Task WriteBytes(IPokeAByteProperty target, byte[] bytesToWrite, bool? freeze);
 }
 
 public delegate Task InstanceProcessingAbort();