+++ src/PokeAByte.Web/MultiplePropertiesUpdater.cs
@@ -6,7 +6,7 @@ namespace PokeAByte.Web;
 
 public static class MultiplePropertiesUpdater
 {
-    public static byte[] ConstructUpdatedBytes(Dictionary<IPokeAByteProperty, string> properties)
+    public static byte[] ConstructUpdatedBytes(Dictionary<IPokeAByteProperty, string> properties, IPokeAByteMapper mapper)
     {
         if (properties.First().Key.Length is null || properties.First().Key.Length < 0)
             throw new ArgumentException("Property length cannot be null or zero.");
@@ -24,13 +24,13 @@ public static class MultiplePropertiesUpdater
             throw new InvalidOperationException("Total bit count exceeds the maximum size of the bit array.");
 
         //We want to fill this new output array with the original value to maintain consistency 
-        var outputBits = new BitArray(properties.First().Key.BytesFromFullValue());
+        var outputBits = new BitArray(properties.First().Key.BytesFromFullValue(mapper));
 
         //Construct the new bitarray using the data from the properties
         foreach (var prop in properties)
         {
             var bitRange = PropertyLogic.ParseBits(prop.Key.Bits);
-            var inputBits = new BitArray(prop.Key.BytesFromValue(prop.Value));
+            var inputBits = new BitArray(prop.Key.BytesFromValue(prop.Value, mapper));
             for (var i = 0; i < bitRange.Length; i++)
             {
                 outputBits[bitRange[i]] = inputBits[i];